DS-1812: move datasources utils to datasources/common
