+++ query/src/datasources/common/line_test.rs
@@ -12,18 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#[cfg(test)]
-mod tests;
+use std::env;
 
-mod fuse_table;
-mod index;
-mod meta_backend;
+use common_exception::Result;
+use pretty_assertions::assert_eq;
 
-pub(crate) mod common;
-pub(crate) mod dal;
-pub(crate) mod example;
-pub(crate) mod local;
-pub(crate) mod remote;
-pub(crate) mod system;
+use crate::datasources::common::count_lines;
 
-pub use meta_backend::MetaBackend;
+#[test]
+fn test_lines_count() -> Result<()> {
+    let file = env::current_dir()?
+        .join("../tests/data/sample.csv")
+        .display()
+        .to_string();
+
+    let lines = count_lines(std::fs::File::open(file.as_str())?)?;
+    assert_eq!(6, lines);
+    Ok(())
+}
