+++ query/src/datasources/common/part_test.rs
@@ -13,17 +13,16 @@
 // limitations under the License.
 
 use common_exception::Result;
+use common_planners::Part;
+use pretty_assertions::assert_eq;
+
+use crate::datasources::common::generate_parts;
 
 #[test]
 fn test_util_generate_parts() -> Result<()> {
-    use common_planners::Part;
-    use pretty_assertions::assert_eq;
-
-    use crate::datasources::Common;
-
     {
         // deal with remainder
-        let ps = Common::generate_parts(0, 3, 11);
+        let ps = generate_parts(0, 3, 11);
 
         assert_eq!(3, ps.len());
         assert_eq!(
@@ -51,7 +50,7 @@ fn test_util_generate_parts() -> Result<()> {
 
     {
         // total is zero
-        let ps = Common::generate_parts(0, 3, 0);
+        let ps = generate_parts(0, 3, 0);
 
         assert_eq!(1, ps.len());
         assert_eq!(
@@ -64,7 +63,7 @@ fn test_util_generate_parts() -> Result<()> {
     }
     {
         // only one part, total < workers
-        let ps = Common::generate_parts(0, 3, 2);
+        let ps = generate_parts(0, 3, 2);
 
         assert_eq!(1, ps.len());
         assert_eq!(
@@ -76,23 +75,5 @@ fn test_util_generate_parts() -> Result<()> {
         );
     }
 
-    Ok(())
-}
-
-#[test]
-fn test_lines_count() -> Result<()> {
-    use std::env;
-
-    use pretty_assertions::assert_eq;
-
-    use crate::datasources::Common;
-
-    let file = env::current_dir()?
-        .join("../tests/data/sample.csv")
-        .display()
-        .to_string();
-
-    let lines = Common::count_lines(std::fs::File::open(file.as_str())?)?;
-    assert_eq!(6, lines);
     Ok(())
 }
